Improve coverage, simplify code.
